style: 관리자페이지 코드리뷰 토대로 리팩토링
