--- src/components/ProjectSubmissionTable.tsx
@@ -1,64 +1,0 @@
-import React from 'react';
-
-interface Submission {
-  id: number;
-  teamName: string;
-  projectName: string;
-  isSubmitted: boolean;
-  likes: number;
-}
-interface ProjectSubmissionTableProps {
-  submissions: Submission[];
-  type: 'project' | 'vote';
-}
-
-const ProjectSubmissionTable: React.FC<ProjectSubmissionTableProps> = ({ submissions, type }) => {
-  return (
-    <section className="mb-8">
-      <h2 className="mb-8 text-2xl font-bold">{type === 'project' ? '프로젝트 등록현황' : '좋아요 랭킹'}</h2>
-      <table className="w-full border-collapse bg-white">
-        <thead className="bg-gray-100">
-          <tr>
-            <th className="w-[5%] border-r border-b border-gray-300 p-2 text-sm">
-              {type === 'project' ? '순번' : '순위'}
-            </th>
-            <th className="w-[20%] border-r border-b border-gray-300 p-2 text-sm">팀명</th>
-            <th className="w-[55%] border-r border-b border-gray-300 p-2 text-sm">작품명</th>
-            <th className="w-[20%] border-b border-gray-300 p-2 text-sm">
-              {type === 'project' ? '제출여부' : '좋아요 수'}
-            </th>
-          </tr>
-        </thead>
-        <tbody>
-          {submissions.map((item: Submission, idx: number) => (
-            <tr key={item.id} className="">
-              <td className="w-[10%] border-r border-b border-gray-300 p-2 py-3 pl-4 text-sm">{idx + 1}</td>
-              <td className="w-[20%] border-r border-b border-gray-300 p-2 py-3 pl-4 text-sm">{item.teamName}</td>
-              <td className="w-[50%] border-r border-b border-gray-300 p-2 py-3 pl-4 text-sm">{item.projectName}</td>
-              {type === 'vote' ? (
-                <td className="w-[20%] border-b border-gray-300 p-2 py-3 pl-4 text-sm">
-                  <div className="flex h-[30px] w-[100px] items-center">
-                    <span>좋아요&nbsp;&nbsp;</span>
-                    <span className="font-bold">{item.likes.toLocaleString()}</span>개
-                  </div>
-                </td>
-              ) : (
-                <td className="w-[20%] border-b border-gray-300 p-2 py-3 pl-4 text-sm">
-                  <div
-                    className={`flex h-[35px] w-[80px] items-center justify-center rounded-md text-center ${
-                      item.isSubmitted ? 'bg-[#005BAA]' : 'bg-[#EC523E]'
-                    } px-2 py-[4px] text-sm text-white`}
-                  >
-                    {item.isSubmitted ? '제출완료' : '미제출'}
-                  </div>
-                </td>
-              )}
-            </tr>
-          ))}
-        </tbody>
-      </table>
-    </section>
-  );
-};
-
-export default ProjectSubmissionTable;
