+++ src/pages/admin/VoteRate.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { PieChart, Pie, Cell } from 'recharts';
 
 interface PieItem {
@@ -6,14 +5,14 @@ interface PieItem {
   value: number;
 }
 
-interface PieData {
+interface Props {
   pieData: PieItem[];
   totalVotes: number;
   participationRate: number;
   pieColors: string[];
 }
 
-const VoteRate: React.FC<PieData> = ({ pieData, totalVotes, participationRate, pieColors }) => {
+const VoteRate = ({ pieData, totalVotes, participationRate, pieColors }: Props) => {
   return (
     <section className="">
       <h2 className="mb-4 text-2xl font-bold">투표 참여율</h2>
@@ -23,7 +22,6 @@ const VoteRate: React.FC<PieData> = ({ pieData, totalVotes, participationRate, p
       >
         <div className="flex justify-center">
           <PieChart width={150} height={150}>
-            {/* 1. 회색 배경 원 전체 */}
             <Pie
               data={[{ name: '배경', value: 100 }]}
               cx="50%"
@@ -35,10 +33,8 @@ const VoteRate: React.FC<PieData> = ({ pieData, totalVotes, participationRate, p
               dataKey="value"
               cornerRadius={0}
             >
-              <Cell fill={pieColors[1]} stroke="none" /> {/* 회색 */}
+              <Cell fill={pieColors[1]} stroke="none" />
             </Pie>
-
-            {/* 2. 초록 참여율 조각만 위에 덧칠 */}
             <Pie
               data={[{ name: '참여', value: participationRate }]}
               cx="50%"
@@ -50,23 +46,8 @@ const VoteRate: React.FC<PieData> = ({ pieData, totalVotes, participationRate, p
               dataKey="value"
               cornerRadius={10}
             >
-              <Cell fill={pieColors[0]} /> {/* 초록 */}
+              <Cell fill={pieColors[0]} />
             </Pie>
-
-            {/* <Pie
-              data={pieData}
-              cx="50%"
-              cy="50%"
-              innerRadius={50}
-              outerRadius={70}
-              dataKey="value"
-              startAngle={-270}
-              endAngle={90}
-            >
-              {pieData.map((entry, index) => (
-                <Cell key={`cell-${index}`} fill={pieColors[index]} />
-              ))}
-            </Pie> */}
           </PieChart>
         </div>
         <p className="-mt-20 text-xl font-bold text-green-600">{participationRate}%</p>
