+++ src/route/AppRoutes.tsx
@@ -8,7 +8,7 @@ import Login from '@pages/auth/Login';
 import Main from '@pages/main/Main';
 import ProjectEditor from '@pages/project-editor/ProjectEditor';
 import ProjectViewer from '@pages/project-viewer/ProjectViewer';
-import AdminDashboard from '@pages/admin/AdminDashboard';
+import AdminPage from '@pages/admin/AdminPage';
 
 const AppRoutes = () =>
   useRoutes([
@@ -46,7 +46,7 @@ const AppRoutes = () =>
             },
             {
               path: 'monitor',
-              element: <AdminDashboard />,
+              element: <AdminPage />,
             },
           ],
         },
